some final primping and minor fixes before merge.
